+++ demos/tileset-sprite/src/application.ts
@@ -2,25 +2,32 @@ import {
   Rect,
   Sprite,
   Vec2,
-  createContext,
-  createInput,
   createTextureLoader,
-  createTimer,
-  inputControlledCamera,
-  pipeline,
   sprite,
 } from '@pixel-craft/engine';
-import Stats from 'stats.js';
+import {
+  Animated,
+  AnimatorSystem,
+  Application,
+  InputCameraSystem,
+  InputSystem,
+  RenderSystem,
+  TimerSystem,
+} from '@pixel-craft/pixel-craft';
 
 export async function application(canvas: HTMLCanvasElement): Promise<void> {
-  const stats = new Stats();
-  document.body.appendChild(stats.dom);
+  const app = await Application.create(canvas);
+  const renderer = new RenderSystem();
+  const input = new InputSystem();
+  const timer = new TimerSystem();
+  const camera = new InputCameraSystem(input, timer);
+  const animator = new AnimatorSystem(timer);
+  await app.addSystems(renderer, input, timer, camera, animator);
 
-  const context = await createContext(canvas);
   const scaling: Vec2.Vec2 = [4, 4];
-  context.camera.zoom(scaling);
+  app.context.camera.zoom(scaling);
 
-  const textureLoader = createTextureLoader(context.device);
+  const textureLoader = createTextureLoader(app.context.device);
 
   const [atlasFloor, atlasCharacters] = await Promise.all([
     textureLoader('assets/0x72_DungeonTilesetII_v1.7/atlas_floor-16x16.png'),
@@ -34,9 +41,6 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
   const tilesX = Math.ceil(canvas.width / tileSize / scaling[0]);
   const tilesY = Math.ceil(canvas.height / tileSize / scaling[1]);
 
-  const renderPass = pipeline(context);
-  const sprites: Sprite[] = [];
-
   const groundTileFrame = {
     x: 0,
     y: 0,
@@ -46,7 +50,7 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
 
   for (let y = 0; y < tilesY; y++) {
     for (let x = 0; x < tilesX; x++) {
-      sprites.push(
+      app.addGameObjects(
         sprite({
           texture: atlasFloor,
           x: x * tileSize,
@@ -69,45 +73,19 @@ export async function application(canvas: HTMLCanvasElement): Promise<void> {
     });
   }
 
-  const hero = sprite({
-    texture: atlasCharacters,
-    x: (tilesX / 2) * tileSize,
-    y: (tilesX / 2 - 4) * tileSize,
-    width: tileSize,
-    height: tileSize * 2,
-    frame: heroFrames[0],
-  });
-  sprites.push(hero);
-
-  const input = createInput();
-  const timer = createTimer();
-  const camera = inputControlledCamera(input, timer, context.camera);
-
-  let heroFrame = 0;
-  const heroSpeed = 5;
-  let heroTimer = 0;
-
-  const draw = function (now: number) {
-    timer.update(now);
-    camera.update();
-    stats.begin();
-
-    hero.frame = heroFrames[heroFrame];
-    heroTimer += timer.deltaTime;
-    if (heroTimer > heroSpeed) {
-      heroTimer = 0;
-      heroFrame++;
-    }
-
-    if (heroFrame >= heroFrames.length) {
-      heroFrame = 0;
-    }
-
-    renderPass(sprites);
-
-    stats.end();
-    requestAnimationFrame(draw);
+  const hero: Sprite & Animated = {
+    ...sprite({
+      texture: atlasCharacters,
+      x: (tilesX / 2) * tileSize,
+      y: (tilesY / 2) * tileSize,
+      width: tileSize,
+      height: tileSize * 2,
+      frame: heroFrames[0],
+    }),
+    animationFrame: 0,
+    animationSpeed: 5,
+    animationTimer: 0,
+    animationFrames: heroFrames,
   };
-
-  draw(performance.now());
+  app.addGameObjects(hero);
 }