+++ packages/engine/src/index.ts
@@ -1,9 +1,4 @@
 /// <reference types="@webgpu/types" />
-export {
-  inputControlledCamera,
-  type InputControlledCamera,
-} from './camera/input-controlled-camera';
-export { createInput, type Input } from './input/input';
 export { createFontLoader } from './loader/font';
 export { createTextureLoader } from './loader/texture';
 export * as Mat4 from './math/mat4';
@@ -11,7 +6,7 @@ export { type Rect } from './math/rect';
 export * as Vec2 from './math/vec2';
 export * as Vec3 from './math/vec3';
 export * as Vec4 from './math/vec4';
-export { createCamera } from './renderer/camera/camera';
+export { type Camera } from './renderer/camera/camera';
 export {
   createContext,
   type WebGPUContext,
@@ -25,4 +20,3 @@ export {
   type TilingSprite,
 } from './sprite/sprite';
 export { canvasText, type CanvasText } from './text/canvas-text';
-export { createTimer, type Timer } from './timer/timer';